requirements.txt
alembic==1.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
mysqlclient==2.2.0
pycparser==2.21
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
